cogview4-6B/requirements.txt
torch --index-url https://download.pytorch.org/whl/cu128
torchvision --index-url https://download.pytorch.org/whl/cu128
transformers
sentencepiece
fastapi
uvicorn
pillow
pydantic
accelerate
diffusers==0.33.1
ninja
protobuf
python-multipart